Create Basic Queue Implementation using Class & Object Concept
I have created a class named queue in which there are 3 methods __init__ for creating the structure of the queue, then there are insert and remove for inserting and removing the data from the queue. I have also created an additional function named operation() which gives a menu of the operations that the user can perform on the queue and shows error for any invalid or wrong entry.